+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - GetContentProtections, GetMimeType, GetCodecs, and GetSegmentTemplate methods for AdaptationSet and Representation
+- ContentProtection elements and name spaces for Marlin DRM and DASH-IF ClearKey
 
 ## [0.10.0] - 2023-05-26